Updated formatting
Hopefully this fixes the formatting issues. I checked out the way another student was doing it; it seems that a line break between the ~~~ and previous text was the only difference between their and my format. That being said, I previewed using github's preview and using another site and both ways (with and without whitespace) rendered correctly, so I don't actually know if my changes actually fixed the formatting.