Added question to adv-wk3
Left in parts of the second question which I think are useful, or at the very
least I want to talk to the rest of the team about first. For the question
I wronte, the skeleton is there. I still need to add the link (and post the
files there). Maybe we want to tell them to use wget and tar.