Update wk6 homework to de-emphasize git hooks
Shifted some content around to place the emphasis on the automated
testing script rather than the git hook aspect.

Instead of telling them this is a git hook, take that complexity out
(seems like lots of students get confused or waste a lot of time here)
and instead just have them install the hook as a last step.