Disable external link checking in htmlproofer
I'm tired of poking at this all the time. The only reason we _actually_
use htmlproofer is is to remind us to release homeworks.

External links generate far too many false positives. We don't have that
many of them, and if one breaks it's not the end of the world.