Update adv-hw-5 to state the grading expectations
Previously we never warned students about the length of five commits as well as
commit message lengths required for full points on part 1 of the assignment.
Now there is a small blurb describing these requirements so things are more
transparent. A link to the grading script is also included for students to
explore. Previously, this script was only linked to in the email sent out with
the grade, so it didn't really help students be more prepared.